+++ spec/features/filter_spec.rb
@@ -14,9 +14,8 @@ RSpec.feature "Service filtering" do
       visit services_path
 
       expect(body).to have_text(root.name)
-      # https://github.com/teamcapybara/capybara/issues/1440#issuecomment-62335948
-      expect(body).to have_text("\u00a0\u00a0#{sub.name}")
-      expect(body).to have_text("\u00a0\u00a0\u00a0\u00a0#{subsub.name}")
+      expect(body).to have_text(sub.name)
+      expect(body).to have_text(subsub.name)
     end
 
     it "shows services from selected research area and sub research areas" do
@@ -29,13 +28,13 @@ RSpec.feature "Service filtering" do
       create(:service, research_areas: [subsub])
       create(:service)
 
-      visit services_path(research_area: root.id)
+      visit services_path(research_areas: [root.id])
       expect(page).to have_selector(".media", count: 3)
 
-      visit services_path(research_area: sub.id)
+      visit services_path(research_areas: [sub.id])
       expect(page).to have_selector(".media", count: 2)
 
-      visit services_path(research_area: subsub.id)
+      visit services_path(research_areas: [subsub.id])
       expect(page).to have_selector(".media", count: 1)
     end
   end